Refactoring improvement
