+++ react/src/js/components/Consonant/CardsCarousel/__tests__/CardsCarousel.spec.js
@@ -20,7 +20,7 @@ import {
     hidePrevButton, 
     hideNextButton, 
     setAriaAttributes 
-} from '../CardsCarouselUtils';
+} from '../CardsCarousel';
 
 global.fetch = jest.fn(() =>
     Promise.resolve({
